refactor. 패키지명 변경(step3->step3to5)
