+++ src/main/java/step3to5/utils/InputValidator.java
@@ -1,4 +1,4 @@
-package step3.utils;
+package step3to5.utils;
 
 import java.util.Scanner;
