+++ src/main/java/step3to5/utils/TestNumberGenerator.java
@@ -1,4 +1,4 @@
-package step3.utils;
+package step3to5.utils;
 
 public class TestNumberGenerator implements NumberGenerator {
