+++ src/test/java/step3to5/domain/CarTest.java
@@ -1,10 +1,10 @@
-package step3.domain;
+package step3to5.domain;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import step3.utils.GameRule;
-import step3.utils.NumberGenerator;
-import step3.utils.TestNumberGenerator;
+import step3to5.utils.GameRule;
+import step3to5.utils.NumberGenerator;
+import step3to5.utils.TestNumberGenerator;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
