+++ src/test/java/step3to5/domain/CarsTest.java
@@ -1,4 +1,4 @@
-package step3.domain;
+package step3to5.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
