+++ src/test/java/step3to5/utils/ConvertToStringFromListTest.java
@@ -1,4 +1,4 @@
-package step3.utils;
+package step3to5.utils;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
