+++ src/test/java/step3to5/utils/GameRuleTest.java
@@ -1,4 +1,4 @@
-package step3.utils;
+package step3to5.utils;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
